Split AST expression operators
